Refactor(parallel coordinates plot): Have parent and child classes
